refactor: :art: improves code reliability by resolving issues reported by the linter and adding instructions for installing the code formatter
