+++ README.md
@@ -1,1 +1,33 @@
-# LinesStructCode
+# Lines Counter
+
+- Versión recomendada de Python: 3.11
+
+## Instalar el entorno (Antes de ejecutar el programa):
+
+1. Crear el entorno de desarrollo:
+
+```bash
+python3 -m venv env
+```
+
+2. Activar el entorno de desarrollo:
+
+En `Windows`:
+
+```bash
+env\Scripts\activate
+```
+
+En `Unix/Linux` o `MacOS`:
+
+```bash
+source env/bin/activate
+```
+
+3. Instalar las dependencias:
+
+```bash
+pip install -r requirements.txt
+
+pre-commit install
+```
