+++ src/test/java/mantenimiento/codecounter/models/JavaFileTest.java
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class JavaFileTest {
+class JavaFileTest {
 
   private Path tempFile;
